r36s: add rg351p support + ensure menumissing.txt is not delected

+++ workspace/r36s/boot/EnableMyMinUI.sh
@@ -13,9 +13,17 @@ if [ ! -e /roms/MyMinUI/r36s/bak ]; then
     cp /usr/bin/emulationstation/emulationstation.sh.es /roms/MyMinUI/r36s/bak/
 fi
 echo "Copy new launcher files"
-sudo cp -vf /roms/MyMinUI/r36s/emulationstation.sh /usr/bin/emulationstation/
-sudo cp -vf /roms/MyMinUI/r36s/emulationstation.sh.ra /usr/bin/emulationstation/
-sudo cp -vf /roms/MyMinUI/r36s/emulationstation.sh.es /usr/bin/emulationstation/
+
+if [ -f "/dev/input/by-path/platform-ff300000.usb-usb-0:1.2:1.0-event-joystick" ]; then
+    # RG351P detected, use its specific launcher
+    sudo cp -vf /roms/MyMinUI/r36s/emulationstation_rg351p.sh /usr/bin/emulationstation/emulationstation.sh
+else
+    # R36S detected, use its specific launcher
+    sudo cp -vf /roms/MyMinUI/r36s/emulationstation.sh /usr/bin/emulationstation/
+    sudo cp -vf /roms/MyMinUI/r36s/emulationstation.sh.ra /usr/bin/emulationstation/
+    sudo cp -vf /roms/MyMinUI/r36s/emulationstation.sh.es /usr/bin/emulationstation/
+fi
+
 echo "Rebooting now..."
 sleep 3
 systemctl reboot
